+++ modules/core/src/service/dynamic-form.service.ts
@@ -1,7 +1,12 @@
 import {Injectable} from "@angular/core";
 import {FormBuilder, FormControl, FormGroup, Validators} from "@angular/forms";
 import {IDynamicFormModel} from "../model/dynamic-form.model";
-import {DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP, DynamicCheckboxGroupModel} from "../model/checkbox/dynamic-checkbox-group.model";
+import {DynamicFormControlModel} from "../model/dynamic-form-control.model";
+import {DYNAMIC_FORM_CONTROL_TYPE_ARRAY} from "../model/dynamic-form-array-control.model";
+import {
+    DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP,
+    DynamicCheckboxGroupModel
+} from "../model/checkbox/dynamic-checkbox-group.model";
 
 @Injectable()
 export class DynamicFormService {
@@ -12,25 +17,30 @@ export class DynamicFormService {
 
         this.formBuilder = formBuilder;
     }
-    
+
     createFormGroup(dynamicFormModel: IDynamicFormModel): FormGroup {
 
         let formGroup = {};
-        
-        dynamicFormModel.items.forEach(controlModel => {
-            
-            if (controlModel.type !== DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP) {
+
+        dynamicFormModel.items.forEach(item => {
+
+            if (item.type === DYNAMIC_FORM_CONTROL_TYPE_ARRAY) {
+
+                // TODO
+
+            } else if (item.type === DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX_GROUP) {
+
+                formGroup[item.id] = this.createFormGroup(<DynamicCheckboxGroupModel> item);
+
+            } else {
+
+                let controlModel = <DynamicFormControlModel<any>> item;
 
                 formGroup[controlModel.id] = new FormControl(
-
                     controlModel.value || null,
                     Validators.compose(controlModel.validators),
                     Validators.composeAsync(controlModel.validatorsAsync)
                 );
-                
-            } else {
-                
-                formGroup[controlModel.id] = this.createFormGroup(<DynamicCheckboxGroupModel> controlModel);
             }
         });
